Viser artikkel og artikkel-rutenett i samme side, felles lasting
