+++ src/components/main/main.tsx
@@ -10,6 +10,7 @@ import NewTermPage from '../new-term-page/new-term-page';
 import { useQueryClient } from '@tanstack/react-query';
 import { fetchTerms } from '../../lib/fetch';
 import { ArticlePage } from '../article-page/article-page';
+import { Article } from '../article-page/article/article';
 
 function Redirect({ to }: any): JSX.Element {
   const navigate = useNavigate();
@@ -28,7 +29,9 @@ const Main = (): JSX.Element => {
       <Header />
       <Routes>
         <Route path="/hjem" element={<Home />} />
-        <Route path="/artikkel/:articleKey" element={<ArticlePage />} />
+        <Route path="/artikkel" element={<ArticlePage />}>
+          <Route path=":articleKey" element={<Article />} />
+        </Route>
         <Route path="/termliste" element={<DictionaryPage />} />
         <Route path="/term/:termId" element={<TermPage />} />
         <Route path="/ny-term" element={<NewTermPage />} />
