+++ src/components/utils/use-articles.ts
@@ -1,0 +1,8 @@
+import { useQuery } from '@tanstack/react-query';
+import { fetchArticles } from '../../lib/fetch';
+
+export const useArticles = () =>
+  useQuery({
+    queryKey: ['articles'],
+    queryFn: fetchArticles,
+  });
